+++ xchange-bitmex/src/main/java/info/bitrich/xchangestream/bitmex/dto/BitmexTrade.java
@@ -6,7 +6,6 @@ import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.Order;
 import org.knowm.xchange.dto.marketdata.Trade;
 
-import java.math.BigDecimal;
 
 /**
  * Created by Lukas Zaoralek on 13.11.17.
@@ -18,8 +17,8 @@ public class BitmexTrade extends BitmexLimitOrder {
     @JsonCreator
     public BitmexTrade(@JsonProperty("symbol") String symbol,
                        @JsonProperty("side") String side,
-                       @JsonProperty("price") BigDecimal price,
-                       @JsonProperty("size") BigDecimal size,
+                       @JsonProperty("price") Double price,
+                       @JsonProperty("size") Double size,
                        @JsonProperty("trdMatchID") String trdMatchID,
                        @JsonProperty("timestamp") String timestamp) {
         super(symbol, "", side, price, size, timestamp);
